+++ EndToEndTests/SampleUseTest.php
@@ -68,9 +68,15 @@ class Epa_EndToEndTests_SampleUseTest extends PHPUnit_Framework_TestCase
 		$login->login('foo', 'bar');
 		$this->assertEquals('last login was foo:bar', $loginEventLogger->getLastLogin());
 
-		# A similar plugin already exists (`Epa\InterfaceToEventNamePlugin`). It adds
-		# event names that are specified inthe doc comments of the interfaces
-		# an event implements. `@event` adds the interface name as an event.
-		# `@eventName someName` adds the specified name as an event.
+		# A similar plugin already exists (`Epa\MetaEventNamePlugin`). It adds
+		# event names that are specified inthe doc comments of the classes/interfaces
+		# an event extends/implements. `@eventname` adds the interface or class name
+		# as an event. `@eventName someName` adds the specified name as an event.
+		#
+		#     /**
+		#      * @eventName Foo 
+		#      */
+		#
+		# This will add the name 'Foo' as an eventname.
 	}
 }