+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Epa Changelog
 =============
 
+0.2.3
+
+	* Plugin allows that eventnames can also be specified in the PHPDoc
+	  of all extended classes (`Epa\MetaEventNamePlugin`). Removed `@event`
+	  so only `@eventname` needs to be used. If no name is specified it uses
+	  the class name.
+
 0.2.2
 
 	* Added a plugin that looks at the doc comments of interfaces that are implemented